move grib parser to own file
